Don't pursue cleanup after buffer kill anymore
Burying a "transient" buffer usually closes it together with the window
if it wasn't meant to reuse the current window anyways and is the ideal
cleanup behaviour.  In case this doesn't work, closing its window works,
too.  Killing it would introduce unforeseeable consequences and is
generally complicating matters, therefore I won't pursue making that
work anymore and will document it in a later commit.